+++ src/main/java/com/javasampleapproach/springrest/mysql/model/records_in_matches/RecordsInMatchesRequest.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class RecordsInMatchesRequest {
     String competition;
-    Integer teamId;
+    Long teamId;
 }
